refactor SearchPageNarrow.tsx remove redundant code
